SQL Output Overhaul
It's now much easier to add SQL output support for existing models..

To add support, just add an SQLProperties property to the class and add
the names of the properties, the sql ouptut names, and the sql types to
it and it'll handle writing the rest of the old sql queries. Reflection
is fun!

Then add the class to the InitStore in Tools_DB.cs, so it passes the
compatibility checks.