Crash fixes, performance improvements, named MTX images
* When doing an XML compare build new MTX and Collections will now
output the image with the name if it is present.
* Upgraded all projects to .NET 4.6 (please enable Server GC as in
app.config.example)
* Disabled the debug flag on SlimDX (this is broken on Windows 10)
* Upgraded HashDictionary to be in line with the latest EasyMYP
dictionary and updated it to .NET 4.6. The hash dictionary has been
rebuilt. Non English and assets that can't be found with the filename
finders will be missing.
* Setting GC to low latency mode while loading Node, Asset or Model
browser.
* Fixed several crashes when previewing ACB files.
* Fixed a crash when selecting multiple files within an ACB file.
* No longer plays audio when selecting multiple files in the tree list
view.
* The model browser nodes key now contains a "new" sub key that shows
new nodes since the last version. This is currently only for nodes and
not mounts and such.
* Fixed a crash when trying to load the model browser after closing the
asset browser.
* Hash dictionary is now reloaded after trying to find new cartel market
images to avoid the asset browser not detecting the new gfx files.
* Added Mastery stat.
* Added several checks to avoid crashes when comparing builds.
* Added absorb co-efficient output for XML output.
* Overall improved performance considerably, must be using server GC as
workstation GC seems to have issues with lots of small object
allocations in .NET 4.6. Memory usage will be considerably higher.